Renaming live-reload-server to live-reload.
